refactor: updated data-context with new state variables and imports
