+++ src/components/index.js
@@ -7,3 +7,7 @@ export { CheckBoxInput } from "./Input/CheckBoxInput/CheckBoxInput";
 export { BadgeButton } from "./Button/BadgeButton";
 export { OutlineButton } from "./Button/OutlineButton";
 export { SolidButton } from "./Button/SolidButton";
+export { CartPriceCard } from "./Cart/CartPriceCard/CartPriceCard";
+export { EmptyCart } from "./Cart/EmptyCart/EmptyCart";
+export { ProductCartCard } from "./Cart/ProductCartCard/ProductCartCard";
+export { WishListCard } from "./WishListCard/WishListCard";
