+++ src/index.js
@@ -3,8 +3,8 @@ import ReactDOM from "react-dom";
 import { makeServer } from "./server";
 import { BrowserRouter } from "react-router-dom";
 import { AppRoutes } from "./routes/AppRoutes";
-import { DataProvider } from "./context/data-provider-context";
-import { AuthProvider } from "./context/auth-provider-context";
+import { DataProvider } from "./context/data-context";
+import { AuthProvider } from "./context/auth-context";
 
 // Call make Server
 makeServer();
